[docs] Refactor examples to have a cmd/ binary
Not sure if I really should be doing this, but I'd like to confirm how
stacktraces behave when doing manual testing, and whether or not a
stackframe occurs in the main package/main function or not triggers
different logic.
