+++ examples/cmd/cli/main.go
@@ -1,0 +1,5 @@
+package main
+
+import "github.com/kinbiko/bugsnag/examples/cli"
+
+func main() { cli.Run() }
