+++ examples/cmd/grpc/client/main.go
@@ -1,0 +1,5 @@
+package main
+
+import "github.com/kinbiko/bugsnag/examples/grpc/client"
+
+func main() { client.Run() }
