+++ examples/cmd/grpc/server/main.go
@@ -1,0 +1,5 @@
+package main
+
+import "github.com/kinbiko/bugsnag/examples/grpc/server"
+
+func main() { server.Run() }
