+++ examples/cmd/nethttp/main.go
@@ -1,0 +1,5 @@
+package main
+
+import "github.com/kinbiko/bugsnag/examples/nethttp"
+
+func main() { nethttp.Run() }
