+++ examples/cmd/panics/main.go
@@ -1,0 +1,5 @@
+package main
+
+import "github.com/kinbiko/bugsnag/examples/panics"
+
+func main() { panics.Run() }
