+++ examples/cmd/sanitizer/augment/main.go
@@ -1,0 +1,5 @@
+package main
+
+import "github.com/kinbiko/bugsnag/examples/sanitizer/augment"
+
+func main() { augment.Run() }
