+++ examples/cmd/sanitizer/grouping/main.go
@@ -1,0 +1,5 @@
+package main
+
+import "github.com/kinbiko/bugsnag/examples/sanitizer/grouping"
+
+func main() { grouping.Run() }
