+++ examples/cmd/sanitizer/sanitize/main.go
@@ -1,0 +1,5 @@
+package main
+
+import "github.com/kinbiko/bugsnag/examples/sanitizer/sanitize"
+
+func main() { sanitize.Run() }
