--- examples/o11y/README.md
@@ -1,3 +1,0 @@
-# O11y
-
-This solution is a tip on how to organise observability (o11y) tools in a single location, a pattern I often use in my services.
