+++ examples/sanitizer/sanitize/sanitize.go
@@ -1,4 +1,4 @@
-package main
+package sanitize
 
 import (
 	"context"
@@ -8,7 +8,7 @@ import (
 	"github.com/kinbiko/bugsnag"
 )
 
-func main() {
+func Run() {
 	n, _ := bugsnag.New(bugsnag.Configuration{
 		APIKey:       os.Getenv("BUGSNAG_API_KEY"),
 		AppVersion:   "1.2.3",
